Update to pull data from Codemash via Ajax.
Also disabled search on the main Sessions page, added date and time
to the session table, and made the Ajax not get called multiple times.

Search on the main sessions page is incompatible with the way the
grouping currently works. It doesn't crash but it does result in weird
and inconsistent sorting behavior. Fuse.js sorts by search relevancy,
but I was searching on the full array of sessions and never sorting
again after that. It should be easy to fix.

Favorited session persistence, favoriting from sessions table,
and a loading screen during the initial Ajax all still need
to be implemented.